# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vahlay Connect — real-time chat platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages (paginated, newest first)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: before
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /messages/search:
    get:
      operationId: searchMessages
      tags: [messages]
      summary: Search messages across user's conversations
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: conversationId
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSearchResults"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        title:
          type: ["string", "null"]
        avatarColor:
          type: string
        status:
          type: string
          enum: [online, away, offline, busy, dnd, brb]

        lastSeenAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, avatarColor, status, createdAt]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        kind:
          type: string
          enum: [direct, group]
        name:
          type: ["string", "null"]
        iconEmoji:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMember"
        lastMessage:
          oneOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        unreadCount:
          type: integer
        pinned:
          type: boolean
        muted:
          type: boolean
      required: [id, kind, createdAt, members, unreadCount, pinned, muted]
    ConversationMember:
      type: object
      properties:
        userId:
          type: integer
        isAdmin:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required: [userId, isAdmin, user]
...
    MessageAttachment:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        size:
          type: integer
        mimeType:
          type: string
      required: [name, url, size, mimeType]
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MessageAttachment"
        replyToId:
          type: ["integer", "null"]
        replyTo:
          oneOf:
            - $ref: "#/components/schemas/MessagePreview"
            - type: "null"
        mentions:
          type: array
          items:
            type: integer
        mentionsEveryone:
          type: boolean
        editedAt:
          type: ["string", "null"]
          format: date-time
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        readBy:
          type: array
          items:
            type: integer
      required:
        - id
        - conversationId
        - authorId
        - author
        - content
        - attachments
        - mentions
        - mentionsEveryone
        - createdAt
        - readBy
    MessagePreview:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
      required: [id, authorId, authorName, content]
    MessageList:
      type: array
      items:
        $ref: "#/components/schemas/Message"
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/MessageAttachment"
        replyToId:
          type: ["integer", "null"]
        clientId:
          type: ["string", "null"]
        mentions:
          type: array
          items:
            type: integer
        mentionsEveryone:
          type: boolean
      required: [content]
...
    EditMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    MessageSearchResult:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/Message"
        conversation:
          $ref: "#/components/schemas/Conversation"
      required: [message, conversation]
    MessageSearchResults:
      type: array
      items:
        $ref: "#/components/schemas/MessageSearchResult"
...