# lib/api-zod/src/generated/types/messagePreview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vahlay Connect — real-time chat platform
 * OpenAPI spec version: 0.1.0
 */

export interface MessagePreview {
  id: number;
  authorId: number;
  authorName: string;
  content: string;
}