# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vahlay Connect — real-time chat platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  title?: string | null;
  avatarColor: string;
  status: UserStatus;
  /** @nullable */
  lastSeenAt?: Date | null;
  createdAt: Date;
}